refactor: consolidate E2E tests into st_tabs_test.py
